refactor: :recycle: Migrating from front-end architecture to feature-based: components, slices, services, and types organized by domain.
Migrating from front-end architecture to feature-based: components, slices, services, and types organized by domain.
