+++ .gitignore
@@ -100,8 +100,6 @@ dist
 # DynamoDB Local files
 .dynamodb/
 
-# GitHub Copilot instructions
-.github/copilot-instructions.md
 
 # TernJS port file
 .tern-port
