+++ src/core/helpers/getEnvironments.ts
