+++ src/features/movies/types/ErrorInterface.ts
