+++ src/features/movies/types/index.ts
@@ -1,0 +1,3 @@
+export * from './MovieInterface';
+export type { MoviesState } from './MovieState';
+export type { StateStorage } from './StateStorage';
