--- src/pages/Details/index.ts
@@ -1,1 +1,0 @@
-export { Details } from './Details';
