--- src/pages/ResultsSearch/index.ts
@@ -1,1 +1,0 @@
-export { ResultsSearch } from './ResultsSearch';
