--- src/pages/index.ts
@@ -1,5 +1,0 @@
-export { Details } from './Details';
-export { Home } from './Home';
-export { NotFount } from './NotFount';
-export { Results } from './Results';
-export { ResultsSearch } from './ResultsSearch';
